refactor: refactoring the unlogged header style
