+++ frontend/src/app/401-page/page.tsx
@@ -1,4 +1,4 @@
-import UnLoggedHeader from '@/components/Headers/UnLoggedHeader'
+import UnLoggedHeader from '@/components/Headers/unlogged-header'
 import Link from 'next/link'
 import { IoMdCloseCircleOutline } from 'react-icons/io'
