+++ frontend/src/app/aboutPage/page.tsx
@@ -1,4 +1,4 @@
-import LoggedHeader from '../../components/Headers/LoggedHeader'
+import LoggedHeader from '../../components/Headers/logged-header'
 
 export default function About() {
   return (
