+++ frontend/src/app/createList/page.tsx
@@ -1,5 +1,5 @@
 import GiftListForm from './(components)/NewListForm/NewList'
-import LoggedHeader from '../../components/Headers/LoggedHeader'
+import LoggedHeader from '../../components/Headers/logged-header'
 
 export default function newList() {
   return (
