+++ frontend/src/app/forgot-password/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 import UserButton from '@/components/Common/buttons/user-button/user-button'
 import InputContainer from '@/components/Common/input-container/input-container'
-import UnLoggedHeader from '@/components/Headers/UnLoggedHeader'
+import UnLoggedHeader from '@/components/Headers/unlogged-header'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { z } from 'zod'
