+++ frontend/src/app/giftsTable/[id]/page.tsx
@@ -1,4 +1,4 @@
-import LoggedHeader from '@/components/Headers/LoggedHeader'
+import LoggedHeader from '@/components/Headers/logged-header'
 import TableSet from './(components)/table-set'
 
 export default function giftTablePage() {
