+++ frontend/src/app/page.tsx
@@ -1,5 +1,5 @@
 import LoginForm from '../components/Forms/LoginForm'
-import UnLoggedHeader from '../components/Headers/UnLoggedHeader'
+import UnLoggedHeader from '../components/Headers/unlogged-header'
 
 export default function signIn() {
   return (
