+++ frontend/src/app/registerPage/page.tsx
@@ -1,5 +1,5 @@
 import RegisterForm from './(components)/RegistrationForms/RegistrationForm'
-import UnLoggedHeader from '../../components/Headers/UnLoggedHeader'
+import UnLoggedHeader from '../../components/Headers/unlogged-header'
 
 export default function Register() {
   return (
